Исправлена ошибка в w2.py

+++ python(foxford)/00/w2.py
@@ -2,8 +2,8 @@ import os
 
 while True:
 	try:
-		a = int(input('введи 1 значение: '))
-		b = int(input('введи 2 значение: '))
+		a = float(input('введи 1 значение: '))
+		b = float(input('введи 2 значение: '))
 		ss = (a + b)/2
 		os.system('clear')
 		print(ss)
